[MOVED] Model classes to the common module
