+++ app/src/main/java/dev/hossain/weatheralert/data/WeatherRepository.kt
@@ -1,6 +1,9 @@
 package dev.hossain.weatheralert.data
 import com.slack.eithernet.ApiResult
 import com.squareup.anvil.annotations.ContributesBinding
+import dev.hossain.weatheralert.datamodel.ForecastData
+import dev.hossain.weatheralert.datamodel.Rain
+import dev.hossain.weatheralert.datamodel.Snow
 import dev.hossain.weatheralert.db.CityForecast
 import dev.hossain.weatheralert.db.CityForecastDao
 import dev.hossain.weatheralert.di.AppScope
