+++ app/src/main/java/dev/hossain/weatheralert/ui/details/WeatherAlertDetails.kt
@@ -53,9 +53,9 @@ import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import dev.hossain.weatheralert.data.CUMULATIVE_DATA_HOURS_24
-import dev.hossain.weatheralert.data.WeatherAlertCategory
 import dev.hossain.weatheralert.data.WeatherService
 import dev.hossain.weatheralert.data.icon
+import dev.hossain.weatheralert.datamodel.WeatherAlertCategory
 import dev.hossain.weatheralert.db.Alert
 import dev.hossain.weatheralert.db.AlertDao
 import dev.hossain.weatheralert.db.City
