+++ data-model/src/test/java/dev/hossain/weatheralert/datamodel/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package dev.hossain.weatheralert.datamodel
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
